Populate the Git Repository
Add 'Specification' to the repo.
Includes a copy of 96board specification.
This serves as a holding place for any specification files the user may have.

Add 'Production_Files' to the repo.
When the user has production file (eg gerber), they should store them in this
folder.

Add 'Mezzanine_Board' to the repo.
Contains all the PCB project files, including libraries and footprints.

Add 'Documentation' to the repo.
Documentation for the user's project should be stored in this folder.

Add 'Test to the repo.
Test plans for the PCB should be stored in here.

Signed-off-by: Sophie Haynes <sophie.haynes@linaro.org>