Power FTDI Independently, Make File Paths Relative, Add Reg. Footprint
Previously the FTDI was powered by the baseboard. This means it would only
be powered up once the baseboard is on - resulting in a few seconds of data
loss. This has been fixed by providing a different power source for the FTDI.

The mezza library path is now a relative path so it should work on other
computers when cloned.

A footprint for the regulator has been added.

Signed-off-by: Sophie Haynes <sophie.haynes@linaro.org>