Clean Up Libraries and Add Footprint Note to Schematic
In the previous version of this project, it required access to libraries.
These were removed, but when looking at the modified files in a text
editor, there were still traces of these libraries. These have now
been removed.

Since a user may wish to use a through hole connector for their mezzanine
board rather than the SMD type this project has, a box graphic and text
has been added to the schematic. This tells the user to remove the component
that is in the box as it is not needed.

Signed-off-by: Sophie Haynes <sophie.haynes@linaro.org>