KiCad template: Add footprints for CE form factors
Add footprints for each of the 96Boards CE form factors; standard,
extended-A and extended-B. The footprint contains the correct size and
position of the mounting holes, the position of each of the expansion
connectors, and the dimensions of the keepout areas. Adding one of these
footprints to a board design will make it much easier to make sure
everything is lined up correctly.

Also update the template PCB to use the standard size footprint instead
of discrete 1pin holes, and add a note about how to swap it for the
extended-A or -B footprint.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>