kicad: Merge extended A/B footprints into one
The extended A/B outlines are supposed to be the same size, but the
template showed extended A as being a smaller board. This happened due
to misreading the 96Boards CE spec. Fixed by deleting the Ext'd A template
and renaming the Ext'd B template to simply 96Boards_CE_outline_extd.