Initial schematics for i2s-mezzanine sound card
These schematics are a subset of a much more complex project I have
been working on, and are probably a good head start on an i2s
mezzanine.

The PCM DI and DO of the 96boards LS port are connected to GPIO0 and
GPIO1 of the pcm1865 respectively. GPIO0 is to be configured as
DOUT2, GPIO1 is to be configured as DIN. This is done by setting
register PCM186X_GPIO1_0_CTRL=0x65.

pcm1865 DOUT(1) is routed to the PCM5142 DIN's.

There is an optional 12.288 MHz crystal that can be populated at
U304, along with C332 and C333 if it is desired for the PCM1865
to be able to operate in MASTER mode. When the PCM1865 is operating
in MASTER mode, it is able to supply SCK to the PCM5142s via GPIO2.

3v3 is supplied by L4931-3.3, which can be switched on and off via
GPIO-K. If being switchable is not desired, do not populate R110,
R111, or Q101, do populate R112 with 0 ohm resistor.

Signed-off-by: Adam Serbinski <adam@serbinski.com>